moved some monaderror syntax to applicative error, added syntax to the Syntax object
