+++ core/src/main/scala/scalaz/syntax/MonadErrorSyntax.scala
@@ -4,18 +4,8 @@ package syntax
 /** Wraps a value `self` and provides methods related to `MonadError` */
 final class MonadErrorOps[F[_], S, A] private[syntax](self: F[A])(implicit val F: MonadError[F, S]) {
   ////
-  final def handleError(f: S => F[A]): F[A] =
-    F.handleError(self)(f)
-
   final def emap[B](f: A => S \/ B): F[B] =
     F.emap(self)(f)
-
-  final def recover(f: S => A): F[A] =
-    F.handleError(self)(s => F.point(f(s)))
-
-  final def attempt: F[S \/ A] =
-    F.handleError(F.map(self)(a => \/.right[S, A](a)))(e => F.point(-\/(e)))
-
   ////
 }
